Add ContactFixture for test setup
